Refactor and rename
Compiler is now more correctly named as CodeGenerator, since it outputs a VirtualFile, i.e., a model of an artifact. Definition is removed, leaving Statement to contrast with Binding, so now it should be easier to see that every lang will have entities and definitions, and in scala these are implemented as bindings and statements of bindings.
