+++ modules/compile/src/main/scala/dc10/schema/FileSchema.scala
@@ -5,4 +5,4 @@ import java.nio.file.Path
 /**
  * The type A is, e.g., a single definition, or a list of defintions.
  */ 
-case class FileDef[A](path: Path, contents: A)
+case class FileSchema[A](path: Path, contents: A)
