+++ modules/io/src/main/scala/dc10/io/FileWriter.scala
@@ -2,7 +2,7 @@ package dc10.io
 
 import cats.effect.Concurrent
 import cats.implicits.*
-import dc10.compiler.Compiler.VirtualFile
+import dc10.compiler.CodeGenerator.VirtualFile
 import fs2.{Stream, text}
 import fs2.io.file.{Files, Path}
