+++ modules/io/src/main/scala/dc10/io/package.scala
@@ -4,8 +4,8 @@ import cats.{FlatMap, Foldable, Monoid}
 import cats.data.StateT
 import cats.effect.Concurrent
 import cats.implicits.*
-import dc10.compiler.{Compiler, Config}
-import dc10.schema.FileDef
+import dc10.compiler.{CodeGenerator, Config}
+import dc10.schema.FileSchema
 import fs2.io.file.{Files, Path}
 
 extension [
@@ -17,10 +17,10 @@ extension [
   def compile: G[L] =
     ast.runEmptyS.foldMapA(Concurrent[G].pure)
 
-extension [F[_]: Concurrent: Files, A](res: F[List[FileDef[A]]])
+extension [F[_]: Concurrent: Files, A](res: F[List[FileSchema[A]]])
   def toFile[V](
     using
-      C: Compiler[A],
+      C: CodeGenerator[A],
       D: Config[V],
   ): F[List[Path]] =
     Concurrent[F].map(res)(C.generate)
