+++ modules/scala/src/main/scala/dc10/scala/compile.scala
@@ -1,0 +1,3 @@
+package dc10.scala
+
+object Hi
