+++ modules/scala/src/main/scala/dc10/scala/ctx/package.scala
@@ -1,9 +1,9 @@
 package dc10.scala.ctx
 
 import cats.{Applicative, Functor}
-import dc10.scala.ast.Definition.Statement
+import dc10.scala.ast.Statement
 import dc10.scala.error.CompileError
-import dc10.schema.FileDef
+import dc10.schema.FileSchema
 
 type ErrorF[A] = Either[List[CompileError], A]
 
@@ -14,9 +14,9 @@ extension [F[_]: Applicative: Functor](ctx: List[Statement])
     // TODO
     Applicative[F].pure(s)
 
-extension [F[_]: Applicative: Functor] (ctx: List[FileDef[List[Statement]]])
-  def ext(s: FileDef[List[Statement]]): F[List[FileDef[List[Statement]]]] =
+extension [F[_]: Applicative: Functor] (ctx: List[FileSchema[List[Statement]]])
+  def ext(s: FileSchema[List[Statement]]): F[List[FileSchema[List[Statement]]]] =
     Functor[F].map(namecheck(s))(ctx :+ _)
-  def namecheck(s: FileDef[List[Statement]]): F[FileDef[List[Statement]]] =
+  def namecheck(s: FileSchema[List[Statement]]): F[FileSchema[List[Statement]]] =
     // TODO
     Applicative[F].pure(s)
