+++ modules/scala/src/main/scala/dc10/scala/ctx/predef/Applications.scala
@@ -5,7 +5,7 @@ import cats.data.StateT
 import dc10.scala.ast.Binding
 import dc10.scala.ast.Binding.Term
 import dc10.scala.ast.Binding.Term.TypeLevel.__
-import dc10.scala.ast.Definition.Statement
+import dc10.scala.ast.Statement
 import dc10.scala.ctx.ErrorF
 
 trait Applications[F[_]]:
