+++ modules/scala/src/main/scala/dc10/scala/ctx/predef/Functions.scala
@@ -4,7 +4,7 @@ import cats.implicits.*
 import cats.data.StateT
 import dc10.scala.ast.Binding
 import dc10.scala.ast.Binding.Term
-import dc10.scala.ast.Definition.Statement
+import dc10.scala.ast.Statement
 import dc10.scala.ctx.ErrorF
 
 trait Functions[F[_]]:
