+++ modules/scala/src/main/scala/dc10/scala/ctx/predef/datatype/ComplexTypes.scala
@@ -5,7 +5,7 @@ import cats.implicits.*
 import dc10.scala.ast.Binding
 import dc10.scala.ast.Binding.{CaseClass, Term}
 import dc10.scala.ast.Binding.Term.TypeLevel.__
-import dc10.scala.ast.Definition.Statement
+import dc10.scala.ast.Statement
 import dc10.scala.error.CompileError
 import dc10.scala.ctx.ErrorF
 import dc10.scala.ctx.ext
