+++ modules/scala/src/main/scala/dc10/scala/ctx/predef/namespace/Packages.scala
@@ -1,11 +1,11 @@
 package dc10.scala.ctx.predef.namespace
 
 import cats.data.StateT
-import dc10.scala.ast.Definition.Statement
-import dc10.scala.ast.Definition.Statement.PackageDef
+import dc10.scala.ast.Statement.PackageDef
 import dc10.scala.ctx.ErrorF
-import dc10.schema.FileDef
+import dc10.schema.FileSchema
 import org.tpolecat.sourcepos.SourcePos
+import dc10.scala.ast.Statement
 
 trait Packages[F[_], G[_]]:
   def PACKAGE[A](nme: String, files: G[A])(using sp: SourcePos): F[A]
@@ -14,7 +14,7 @@ object Packages:
 
   trait Mixins extends Packages[
     [A] =>> StateT[ErrorF, List[PackageDef], A],
-    [A] =>> StateT[ErrorF, List[FileDef[List[Statement]]], A],
+    [A] =>> StateT[ErrorF, List[FileSchema[List[Statement]]], A],
   ]:
-    def PACKAGE[A](nme: String, files: StateT[ErrorF, List[FileDef[List[Statement]]], A])(using sp: SourcePos): StateT[ErrorF, List[PackageDef], A] =
+    def PACKAGE[A](nme: String, files: StateT[ErrorF, List[FileSchema[List[Statement]]], A])(using sp: SourcePos): StateT[ErrorF, List[PackageDef], A] =
       ???
